change to right project layout.
change to right project layout.
