--- gs-accessing-data-rest/src/main/java/Application.java
@@ -1,159 +1,0 @@
-package hello;
-
-
-import java.util.Arrays;
-import java.util.Collection;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.CommandLineRunner;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.*;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
-
-import java.util.Arrays;
-
-import javax.xml.ws.RequestWrapper;
-
-import hello.bookmarks.Account;
-import hello.bookmarks.AccountRepository;
-import hello.bookmarks.Bookmark;
-import hello.bookmarks.BookmarkRepository;
-
-/*@Configuration
-@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
-@ComponentScan*/
-@SpringBootApplication
-public class Application {
-
- 	@Bean
-	CommandLineRunner init(AccountRepository accountRepository,
-						   BookmarkRepository bookmarkRepository) {
-		return (evt) -> Arrays.asList(
-				"jhoeller,dsyer,pwebb,ogierke,rwinch,mfisher,mpollack,jlong".split(","))
-				.forEach(
-						a -> {
-							Account account = accountRepository.save(new Account(a,
-									"password"));
-							bookmarkRepository.save(new Bookmark(account,
-									"http://bookmark.com/1/" + a, "A description"));
-							bookmarkRepository.save(new Bookmark(account,
-									"http://bookmark.com/2/" + a, "A description"));
-						});
-	}
-
-	public static void main(String[] args) {
-		SpringApplication.run(Application.class, args);
-	}
-}
-
-@RestController
-@RequestMapping("/{userId}/bookmarks/{param}")
-class BookmarkRestControllerTest {
-
-    private final BookmarkRepository bookmarkRepository;
-
-    private final AccountRepository accountRepository;
-
-
-    @RequestMapping(method = RequestMethod.POST)
-    ResponseEntity<?> add(@PathVariable String userId, @PathVariable String param, @RequestBody Bookmark input) {
-        System.out.println(param);
-        return null;
-    }
-    @RequestMapping(method = RequestMethod.GET)
-    Collection<Bookmark> readBookmarks(@PathVariable String userId) {
-        return this.bookmarkRepository.findByAccountUsername(userId);
-    }
-    @Autowired
-    BookmarkRestControllerTest(BookmarkRepository bookmarkRepository,
-                               AccountRepository accountRepository) {
-        this.bookmarkRepository = bookmarkRepository;
-        this.accountRepository = accountRepository;
-        System.out.println(bookmarkRepository);
-    }
-
-    private void validateUser(String userId) {
-        this.accountRepository.findByUsername(userId).orElseThrow(
-                () -> new UserNotFoundException(userId));
-    }
-}
-
-@RestController
-@RequestMapping("/{userId}/bookmarks")
-class BookmarkRestController {
-
-	private final BookmarkRepository bookmarkRepository;
-
-	private final AccountRepository accountRepository;
-
-	@RequestMapping(method = RequestMethod.POST)
-    ResponseEntity<?> add(@PathVariable String userId, @RequestBody Bookmark input) {
-		this.validateUser(userId);
-		/*
-		 curl test cmd:
-		    curl -i -H "Content-Type: application/json" -X POST  --data '{"uri":"http://news.sina.com.cn","description":"新浪新闻首页"}'  http://127.0.0.1:8080/dsyer/bookmarks
-
-		 */
-		return this.accountRepository
-				.findByUsername(userId)
-				.map(account -> {
-					Bookmark result = bookmarkRepository.save(new Bookmark(account,
-							input.uri, input.description));
-
-					HttpHeaders httpHeaders = new HttpHeaders();
-					httpHeaders.setLocation(ServletUriComponentsBuilder
-							.fromCurrentRequest().path("/{id}")
-							.buildAndExpand(result.getId()).toUri());
-					return new ResponseEntity<>(null, httpHeaders, HttpStatus.CREATED);
-				}).get();
-
-	}
-
-	@RequestMapping(value = "/{bookmarkId}", method = RequestMethod.GET)
-	Bookmark readBookmark(@PathVariable String userId, @PathVariable Long bookmarkId) {
-		this.validateUser(userId);
-		return this.bookmarkRepository.findOne(bookmarkId);
-	}
-
-	@RequestMapping(method = RequestMethod.GET)
-	Collection<Bookmark> readBookmarks(@PathVariable String userId) {
-		this.validateUser(userId);
-		return this.bookmarkRepository.findByAccountUsername(userId);
-	}
-
-	@Autowired
-	BookmarkRestController(BookmarkRepository bookmarkRepository,
-						   AccountRepository accountRepository) {
-		this.bookmarkRepository = bookmarkRepository;
-		this.accountRepository = accountRepository;
-	}
-
-	private void validateUser(String userId) {
-		this.accountRepository.findByUsername(userId).orElseThrow(
-				() -> new UserNotFoundException(userId));
-	}
-}
-
-@ResponseStatus(HttpStatus.NOT_FOUND)
-class UserNotFoundException extends RuntimeException {
-
-	public UserNotFoundException(String userId) {
-		super("could not find user '" + userId + "'.");
-	}
-}
-
-
