--- gs-accessing-data-rest/src/main/java/hello/bookmarks/AccountRepository.java
@@ -1,10 +1,0 @@
-package hello.bookmarks;
-
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface AccountRepository extends JpaRepository<Account, Long> {
-    Optional<Account> findByUsername(String username);
-}
