+++ gs-accessing-data-rest/src/main/java/restful/Application.java
@@ -1,0 +1,42 @@
+package restful;
+
+import java.util.Arrays;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Bean;
+
+import restful.hello.bookmarks.Account;
+import restful.hello.bookmarks.AccountRepository;
+import restful.hello.bookmarks.Bookmark;
+import restful.hello.bookmarks.BookmarkRepository;
+
+/*@Configuration
+@EnableAutoConfiguration(exclude={DataSourceAutoConfiguration.class})
+@ComponentScan*/
+@SpringBootApplication
+public class Application {
+
+ 	@Bean
+	CommandLineRunner init(AccountRepository accountRepository,
+						   BookmarkRepository bookmarkRepository) {
+		return (evt) -> Arrays.asList(
+				"jhoeller,dsyer,pwebb,ogierke,rwinch,mfisher,mpollack,jlong".split(","))
+				.forEach(
+						a -> {
+							Account account = accountRepository.save(new Account(a,
+									"password"));
+							bookmarkRepository.save(new Bookmark(account,
+									"http://bookmark.com/1/" + a, "A description"));
+							bookmarkRepository.save(new Bookmark(account,
+									"http://bookmark.com/2/" + a, "A description"));
+						});
+	}
+
+	public static void main(String[] args) {
+		SpringApplication.run(Application.class, args);
+	}
+}
+
+
+
