+++ gs-accessing-data-rest/src/main/java/restful/hello/Person.java
@@ -1,4 +1,4 @@
-package hello;
+package restful.hello;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
