+++ gs-accessing-data-rest/src/main/java/restful/hello/PersonRepository.java
@@ -1,4 +1,4 @@
-package hello;
+package restful.hello;
 
 import java.util.List;
